remove gpkg adapter
