+++ spatialconnect/src/main/java/com/boundlessgeo/spatialconnect/tiles/GpkgRasterSource.java
@@ -1,13 +1,11 @@
 package com.boundlessgeo.spatialconnect.tiles;
 
-import com.boundlessgeo.spatialconnect.dataAdapter.GeoPackageAdapter;
 import com.boundlessgeo.spatialconnect.geometries.SCPolygon;
 import com.boundlessgeo.spatialconnect.stores.GeoPackageStore;
 import com.boundlessgeo.spatialconnect.stores.SCRasterStore;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.model.TileOverlayOptions;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -21,7 +19,7 @@ public class GpkgRasterSource implements SCRasterStore {
 
     @Override
     public void overlayFromLayer(String layerName, GoogleMap map) {
-        Map<String, SCGpkgTileSource> tileSources = ((GeoPackageAdapter) gpkgStore.getAdapter()).getTileSources();
+        Map<String, SCGpkgTileSource> tileSources = gpkgStore.getTileSources();
         if (tileSources.size() > 0) {
             if (tileSources.keySet().contains(layerName)) {
                  map.addTileOverlay(
